requirements.txt
Django==2.2
djangorestframework==3.9.2
djangorestframework-stubs==0.4.2
gunicorn==19.9.0
psycopg2==2.8
python-dotenv==0.10.1
pytz==2018.9
sqlparse==0.3.0

